requirements.txt
streamlit
pandas
scikit-learn
matplotlib
psycopg2
sqlalchemy
python-dotenv
